feat: architecture for deployment on pypi
