--- llmax/clients/__init__.py
@@ -1,7 +1,0 @@
-"""The client object."""
-
-from .multi_ai_client import MultiAIClient
-
-__all__ = [
-    "MultiAIClient",
-]
