--- llmax/external_clients/whisper.py
@@ -1,1 +1,0 @@
-"""Whisper client."""
