--- llmax/messages/message.py
@@ -1,5 +1,0 @@
-"""Message type."""
-
-from typing import Any
-
-Message = Any
