--- llmax/messages/messages.py
@@ -1,5 +1,0 @@
-"""Messages type."""
-
-from .message import Message
-
-Messages = list[Message]
