--- llmax/usage/__init__.py
@@ -1,5 +1,0 @@
-"""Usage package."""
-
-from .usage import ModelUsage
-
-__all__ = ["ModelUsage"]
