--- llmax/utils/__init__.py
@@ -1,5 +1,0 @@
-"""Package with the logger."""
-
-from .logger import logger
-
-__all__ = ["logger"]
