+++ src/llmax/clients/__init__.py
@@ -1,0 +1,7 @@
+"""The client object."""
+
+from .multi_ai_client import MultiAIClient
+
+__all__ = [
+    "MultiAIClient",
+]
