+++ src/llmax/external_clients/__init__.py
@@ -1,0 +1,8 @@
+"""Clients, with async and sync."""
+
+from .clients import get_aclient, get_client
+
+__all__ = [
+    "get_aclient",
+    "get_client",
+]
