+++ src/llmax/external_clients/whisper.py
@@ -1,0 +1,1 @@
+"""Whisper client."""
