+++ src/llmax/messages/__init__.py
@@ -1,0 +1,6 @@
+"""Message and Messages types."""
+
+from .message import Message
+from .messages import Messages
+
+__all__ = ["Message", "Messages"]
