+++ src/llmax/messages/message.py
@@ -1,0 +1,5 @@
+"""Message type."""
+
+from typing import Any
+
+Message = Any
