+++ src/llmax/messages/messages.py
@@ -1,0 +1,5 @@
+"""Messages type."""
+
+from .message import Message
+
+Messages = list[Message]
