+++ src/llmax/models/__init__.py
@@ -1,0 +1,19 @@
+"""This module provides LLM and embedder utilities."""
+
+from .deployment import Deployment
+from .fake import fake_llm
+from .models import COHERE_MODELS, MISTRAL_MODELS, MODELS, OPENAI_MODELS, Model
+from .providers import PROVIDERS, Provider
+
+__all__ = [
+    "COHERE_MODELS",
+    "MISTRAL_MODELS",
+    "MODELS",
+    "OPENAI_MODELS",
+    "PROVIDERS",
+    "Deployment",
+    "Deployment",
+    "Model",
+    "Provider",
+    "fake_llm",
+]
