+++ src/llmax/usage/__init__.py
@@ -1,0 +1,5 @@
+"""Usage package."""
+
+from .usage import ModelUsage
+
+__all__ = ["ModelUsage"]
