+++ src/llmax/utils/logger.py
@@ -1,0 +1,11 @@
+"""Logger module."""
+
+import sys
+
+from loguru import logger
+
+logger.add(
+    sys.stderr,
+    format="{time} {level} {message}",
+    level="INFO",
+)
